Fix typo 👍👍👍👍👍👍👍👍👍👍
This is a commit that was made in order to improve code quality and legibility. 
You know, it is important not to any mistake, especially when you try to write in your native language.

Despite the common lament concerning the “dumbing down” of culture and the arts, we read and write more than ever before. While it is true that people are reading online articles and blogs more than they used to, people are still digesting the classics of English literature as much as they ever were.

It is also true that we increasingly rely on an economy based on knowledge, understanding and information. Good writing is essential to proper communication no matter what tool we use.

The main purpose of writing well is to explain and persuade. Regardless of what field you are in, you can go further in your career if you can explain your ideas well. Often, your writing is the first impression that a prospective client or employer gets of you, whether it is a perfectly crafted email or a flawless cover letter.

Writing well makes you appear more professional and intelligent and you have a better chance of persuading people – to hire you, to buy what you are selling or to implement your ideas.

Furthermore, being able to write well enables you to consolidate and organise your ideas better, build arguments clearly and logically and actually think better. This makes you more confident about expressing your thoughts and can improve your verbal communications skills.

Finally, good writing is about communicating a message well and this requires focus and an engaging style. A good writer can tailor his message for the intended audience – whether it is a press release, a formal business proposal, or a twitter post. A wide range of skills is required to have a successful career. Writing well is a skill that can complement and enhance all those other skills.

Are you in need of a good writer? The Global Services Exchange run by blur Group can deliver Content Services to the highest standard with our crowd of thousands of excellent writers.